# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: KODE Printer API
  version: 1.0.0
  description: |
    Central print, scan and audit for KODE Sports Club.

    Every non-2xx body is the same envelope (`Error` below) and carries a
    `requestId` that also appears in the `X-Request-Id` response header — it is
    the join key between a user's complaint and the logs.

    Lists are keyset-paginated: pass the `nextCursor` from the previous page as
    `cursor`. Offset pagination is not offered at any endpoint.
  license:
    name: Internal, confidential — KODE Sports Club · Technology
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [users]
      summary: List accounts
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: role, in: query, schema: { type: string, enum: [admin, user] } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        '200':
          description: A page of accounts.
    post:
      tags: [users]
      summary: Create an account
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /users/{id}:
    get:
      tags: [users]
      summary: One account
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: The account.
    put:
      tags: [users]
      summary: Update an account
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Updated.
        '409':
          description: Refused — this is the only administrator.
    delete:
      tags: [users]
      summary: Delete an account
      description: Job history survives; the username was snapshotted when each job was created.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204':
          description: Deleted.
  /users/{id}/active:
    put:
      tags: [users]
      summary: Enable or disable an account
      description: Disabling ends that account's sessions immediately.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Updated.
  /users/{id}/password:
    put:
      tags: [users]
      summary: Set an account's password
      description: Ends every session for that account.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Set.
  /users/{id}/printers:
    put:
      tags: [users]
      summary: Replace an account's printer grants
      description: The only write path to printer permissions. The whole set is replaced, so the audit row holds before and after.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Updated.
...
  /users/{id}/setup-link:
    post:
      tags: [users]
      summary: Mint a set-password or reset link
      description: |
        Returns the whole URL exactly once, for an administrator to copy and
        send. Only its hash is stored, so no route can show it again. Minting a
        link revokes whatever was outstanding for that person; a reset link also
        ends every session they have.
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                purpose: { type: string, enum: [setup, reset], default: setup }
      responses:
        '201':
          description: The link.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SetupLink' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    collectorKey:
      type: http
      scheme: bearer
      description: A per-collector API key. Never valid on an administrative route.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    Cursor:
      name: cursor
      in: query
      description: Opaque keyset cursor from the previous page.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
  responses:
    Error:
      description: The standard error envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
              description: Machine-readable code from the error taxonomy.
            message:
              type: string
              description: Written for the person reading it, not for a log.
            details:
              type: object
              additionalProperties: true
              description: Never contains a stack trace, a SQL string or a secret.
            requestId:
              type: string
...
    User:
      type: object
      description: No password field exists on this shape, in either direction.
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        role: { type: string, enum: [admin, user] }
        department:
          type: [string, 'null']
          description: Reporting only. It does not influence access.
        zoneId: { type: [integer, 'null'] }
        mustChangePassword: { type: boolean }
        isActive: { type: boolean }
        hasPassword:
          type: boolean
          description: False while the account exists but nobody has chosen a password for it yet.
        setupLinkExpiresAt:
          type: [string, 'null']
          description: When the outstanding set-password link stops working, or null when there is none.
        printerIds:
          type: array
          items: { type: integer }
          description: The grants, which are the sole source of printer permission.
...
    SetupLink:
      type: object
      description: Returned once, at creation. Only the hash is kept.
      properties:
        url: { type: string }
        purpose: { type: string, enum: [setup, reset] }
        expiresAt: { type: string }
        user:
          type: object
          properties:
            id: { type: integer }
            username: { type: string }
            displayName: { type: [string, 'null'] }
...